+++ src/firmware/SYSCLK/sysclk.h
@@ -1,8 +1,8 @@
 #ifndef _SYSCLK_H_
 #define _SYSCLK_H_
 
-void idle_hours(float hours);
-void idle_mins(float mins);
+void idle_hours(int hours);
+void idle_mins(int mins);
 
 #endif //_SYSCLK_H_
